Other small fixes

+++ trunk/src/kademlia/kademlia/Kademlia.cpp
@@ -258,7 +258,7 @@ uint32 CKademlia::GetKademliaUsers(void)
 	return 0;
 }
 
-uint32 CKademlia::GetKademliaFiles(void)
+size_t CKademlia::GetKademliaFiles(void)
 {
 	if( instance && instance->m_prefs ) {
 		return instance->m_prefs->GetKademliaFiles();
